backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.3.0
numpy>=1.26.4
ffmpeg-python>=0.2.0
aiofiles>=23.2.1
sse-starlette>=2.1.0
